escaperai-api/requirements.txt
# Core dependencies
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0

# Authentication & Security
PyJWT==2.11.0
flask-jwt-extended==4.5.2
werkzeug==2.3.7

# Database
alembic==1.12.0

# Utilities
marshmallow==3.20.1
requests==2.31.0